orchestrator/requirements.txt
anthropic>=0.52.0
pyautogui>=0.9.54
Pillow>=10.0.0
